Move debug out of api
