+++ src/instantiate.ts
@@ -3,8 +3,8 @@ import * as vscode from "vscode";
 import { GlobalConfiguration, onCodeForIBMiConfigurationChange } from "./api/Configuration";
 import Instance from "./Instance";
 import { Terminal } from './views/Terminal';
-import { getDebugServiceDetails } from './api/debug/config';
-import { debugPTFInstalled, isDebugEngineRunning } from './api/debug/server';
+import { getDebugServiceDetails } from './debug/config';
+import { debugPTFInstalled, isDebugEngineRunning } from './debug/server';
 import { setupGitEventHandler } from './filesystems/local/git';
 import { registerActionsCommands } from './commands/actions';
 import { registerCompareCommands } from './commands/compare';
