+++ src/testing/debug.ts
@@ -1,6 +1,6 @@
 import assert from "assert";
 import { TestSuite } from ".";
-import { getJavaHome } from "../api/debug/config";
+import { getJavaHome } from "../debug/config";
 import { instance } from "../instantiate";
 
 export const DebugSuite: TestSuite = {
